Add functions to read for written numbers
Part 2 of the day-1 advent of code was to determine the calibration
value for a given line of text in provided input. The calibration value
is an integer consisting of the first number found when searching the
text from the left concatenated with the first number found when
searching from the right resulting in a 2 digit integer. The catch with
part 2 of the challenge was that a number could either be a spelled out
number or an actual digit.

The key to solving this was writing a function to search from the left,
another to search from the right, and one to build a list of substrings
to be searched for each instance where a numerical digit hadn't yet been
found.

The solution to part 2 of the challenge is 54925