Add functions to solve challenge
Input was obtained by using curl on the CLI with the advent of code
website and the corresponding sesion-id cookie and saved to the
input.txt file.

The main feature of the program is the solution_dict.
Each line of input is read in and parsed - the game number is added
as the key and the corresponding string of rounds to be played is
evaluated using a series of functions. The string of rounds is passed
to build_list_of_dicts function which in splits the string into a list
of single rounds and passes each round to the build_dict function. Each
round is built into a dict and appended to list_of_dicts which is the
key value corresonding to a given game.

After the solutions_dict is built, each round in a particular game is
evaluated for get_round_eligibility and if it passes all tests, its
id is added to the winning_ids list to be summed up at the end for the
solution to the challenge.