Add functions to calculate minimum round values and powers
Part 1 of the challenge handled the ground work for reading input and
storing data it for analysis in hashmap and/or list. In part 2 I added
a function to find the minimum possible round, store it in a hashmap,
and append the minimum hashmap for that round to a list of hashmaps.
This involved tweaking the manage_game() function to evaluate each
round against the current minimum and adjust values accordingly.

From there, I read the min_round_dicts list of dicts into the
get_powers() function and appended the corresponding power to the powers
list.

The solution to part 2 of day 2 is 72596