Add part_1.py as separate program for day 1
Write program to solve day-1 part 1 of adventofcode
The task was to read input from adventofcode.com, parse numerical values
from each string in the input, and submit the sum of each two digit
value that was parsed.

The input is a list of strings that has a mix of letters and numeric
characters. Input was obtained using curl in the CLI, redirecting stdout
to the file "input.txt" and then opening input.txt in to read in the
lines within the day_1.py program.

The solution was 55172.