Add functions to solve the day-3 challenge
Read in input and build a 2d-array.
The get_boundaries() function takes in current array index and
calculates min and max row and column values, respectively, for the
search grid. The function is then passed to the check_neighbor()
function which searches the grid based on the boundaries and returns
true is a current value is both numeric and adjacent to a symbol.