Write functions to solve advent of code day 4
Read input, split into hashmap with key=game_number : value=2d-list
where list[0] is a list of winning_numbers and list[1] is list of
current game numbers.

populate_games_played(row) function reads in a row as an argument,
splits first on ":" to isolate game number, then on "|" to isolate the
winning numbers from the current game numbers and finally populates
the games_played dictionary.

calculate_score() function reads each value in the games_played
dictionary, compares the elements in each row of the 2-d array
stored as the dictionary value for each game, and then populates
the scores list with the corresponding score. The first match makes the card worth one point and each match after the
first doubles the point value of that card.

The answer to part 1 day 4 is 21088.