Read input and store critical values
The current program reads input from the example_input.txt file, splits
line of text, and stores the data accordingly.

The seed numbers are stored in the seeds variable.
The mapping data is stored in the maps dictionary with
keys = (string) name of map : values = 2-d list of 3 set numbers.

The map_source_to_destination function reads in the values from
the maps dictionary and returns a dictionary mapping all source numbers
0 - 100 to their corresponding destination numbers per the map data.