Write program to solve day-6 of AOC 2023
The program reads in text input and splits it.
Time:      7  15   30
Distance:  9  40  200

The get_number_solutions() function takas as arguments a time and a
corresponding record_distance value. The function loops through different
game possibilities and returns an integer for the amount of solutions
that exceed the record_distance. This is repeated for each
(time, distance) pair and appended to the number_solutions list.
The solutions is the product of each element in the number_solutions
list.

The solution for day-6 part1 is 170000.